Add files via upload
This is a client server Messenger project made with Java that holds features like private chat and group chats. It is also connected to a database which holds the Sign up information as well as private and public chats. Features like uploading profile picture and viewing informations are also added